# idm/generated/openapi/src/main/java/com/onyxdb/idm/generated/openapi/apis/ServicesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.onyxdb.idm.generated.openapi.apis;

import com.onyxdb.idm.generated.openapi.models.BadRequestResponse;
import com.onyxdb.idm.generated.openapi.models.NotFoundResponse;
import com.onyxdb.idm.generated.openapi.models.ServiceDTO;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "Services", description = "the Services API")
public interface ServicesApi {

    /**
     * POST /api/v1/services : Create a new service
     *
     * @param serviceDTO  (required)
     * @return Created (status code 201)
     *         or Bad Request (status code 400)
     */
    @Operation(
        operationId = "createService",
        summary = "Create a new service",
        tags = { "Services" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ServiceDTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/services",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ServiceDTO> createService(
        @Parameter(name = "ServiceDTO", description = "", required = true) @Valid @RequestBody ServiceDTO serviceDTO
    );


    /**
     * DELETE /api/v1/services/{serviceId} : Delete a service by ID
     *
     * @param serviceId  (required)
     * @return No Content (status code 204)
     *         or Not Found (status code 404)
     *         or Bad Request (status code 400)
     */
    @Operation(
        operationId = "deleteService",
        summary = "Delete a service by ID",
        tags = { "Services" },
        responses = {
            @ApiResponse(responseCode = "204", description = "No Content"),
            @ApiResponse(responseCode = "404", description = "Not Found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = NotFoundResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/v1/services/{serviceId}",
        produces = { "application/json" }
    )
    
    ResponseEntity<Void> deleteService(
        @Parameter(name = "serviceId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("serviceId") UUID serviceId
    );


    /**
     * GET /api/v1/services : Get all services
     *
     * @return OK (status code 200)
     *         or Bad Request (status code 400)
     */
    @Operation(
        operationId = "getAllServices",
        summary = "Get all services",
        tags = { "Services" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ServiceDTO.class)))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/services",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<ServiceDTO>> getAllServices(
        
    );


    /**
     * GET /api/v1/services/{serviceId} : Get a service by ID
     *
     * @param serviceId  (required)
     * @return OK (status code 200)
     *         or Not Found (status code 404)
     *         or Bad Request (status code 400)
     */
    @Operation(
        operationId = "getServiceById",
        summary = "Get a service by ID",
        tags = { "Services" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ServiceDTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = NotFoundResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/services/{serviceId}",
        produces = { "application/json" }
    )
    
    ResponseEntity<ServiceDTO> getServiceById(
        @Parameter(name = "serviceId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("serviceId") UUID serviceId
    );


    /**
     * PUT /api/v1/services/{serviceId} : Update a service by ID
     *
     * @param serviceId  (required)
     * @param serviceDTO  (required)
     * @return OK (status code 200)
     *         or Not Found (status code 404)
     *         or Bad Request (status code 400)
     */
    @Operation(
        operationId = "updateService",
        summary = "Update a service by ID",
        tags = { "Services" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ServiceDTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = NotFoundResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/v1/services/{serviceId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ServiceDTO> updateService(
        @Parameter(name = "serviceId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("serviceId") UUID serviceId,
        @Parameter(name = "ServiceDTO", description = "", required = true) @Valid @RequestBody ServiceDTO serviceDTO
    );

}